使用MyBatis Generator，完成调整以及分页

+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE question
 (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id LONG AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(50),
     description TEXT,
     gmt_create BIGINT,
